Fix condition in `Filter.Contains`
By construction of a Bloom filter, _all_ the bit positions of
an item must be 1 if the item is in the set. The current code
assumes that an item might be in the set if _any_ of its bit
positions are 1, ending up in a filter that is still correct
but vastly less specific.